pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kaggle-mcp-research"
version = "1.0.0"
description = "Full end-to-end Kaggle + HuggingFace MCP server for VS Code Copilot — upload research ZIPs, auto-design novel architectures, run & monitor Kaggle kernels, and retrieve results."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
  { name = "Vinayak Bhatia", email = "ntpjc2vinayak@gmail.com" }
]
keywords = ["mcp", "kaggle", "huggingface", "research", "copilot", "ai", "machine-learning"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "fastmcp>=2.1.0",
  "mcp>=1.6.0",
  "requests>=2.32.0",
  "httpx>=0.27.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "nbformat>=5.9.0",
  "python-dotenv>=1.0.0",
  "rich>=13.0.0",
  "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
  "black>=23.0.0",
  "ruff>=0.1.0",
  "mypy>=1.5.0",
]

[project.scripts]
kaggle-mcp = "kaggle_mcp.server:main"

[project.urls]
Homepage = "https://github.com/vvinayakkk/kaggle-mcp-research"
Documentation = "https://github.com/vvinayakkk/kaggle-mcp-research#readme"
Repository = "https://github.com/vvinayakkk/kaggle-mcp-research.git"
"Bug Tracker" = "https://github.com/vvinayakkk/kaggle-mcp-research/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/kaggle_mcp"]

[tool.hatch.version]
path = "src/kaggle_mcp/__init__.py"

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
